pyproject.toml
[tool.poetry]
name = "drf-ii"
version = "0.1.0"
description = "Working with 'Order' and 'Produc'"
authors = ["Ezequias Silva Santos <phd.esantos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.15"
pytest-django = "^4.8.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
drf-spectacular = "^0.27.2"
psycopg = "^3.2.1"


[tool.poetry.dev-dependencies]
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
